+++ README.md
@@ -49,7 +49,7 @@ pypi: https://pypi.org/project/aio.core
 
 #### [aio.run.runner](aio.run.runner)
 
-version: 0.2.0
+version: 0.2.1
 
 pypi: https://pypi.org/project/aio.run.runner
