+++ aio.run.runner/VERSION
@@ -1,1 +1,1 @@
-0.2.0
+0.2.1
