+++ aio.run.runner/aio/run/runner/decorators.py
@@ -1,3 +1,4 @@
+import inspect
 from functools import wraps
 from typing import Callable, Optional, Tuple, Type, Union
 
@@ -17,36 +18,61 @@ def catches(
     class MyRunner(runner.Runner):
 
         @runner.catches((MyError, MyOtherError))
-        async def run(self):
+        def run(self):
             self.myrun()
     ```
+
+    Can work with `async` methods too.
     """
 
     def wrapper(fun: Callable) -> Callable:
 
         @wraps(fun)
-        async def wrapped(self, *args, **kwargs) -> Optional[int]:
+        def wrapped(self, *args, **kwargs) -> Optional[int]:
+            try:
+                return fun(self, *args, **kwargs)
+            except errors as e:
+                self.log.error(str(e) or repr(e))
+                return 1
+
+        @wraps(fun)
+        async def async_wrapped(self, *args, **kwargs) -> Optional[int]:
             try:
                 return await fun(self, *args, **kwargs)
             except errors as e:
                 self.log.error(str(e) or repr(e))
                 return 1
 
+        wrapped_fun = (
+            async_wrapped
+            if inspect.iscoroutinefunction(fun)
+            else wrapped)
+
         # mypy doesnt trust `@wraps` to give back a `__wrapped__` object so we
         # need to code defensively here
-        wrapping = getattr(wrapped, "__wrapped__", None)
+        wrapping = getattr(wrapped_fun, "__wrapped__", None)
         if wrapping:
             setattr(wrapping, "__catches__", errors)
-        return wrapped
+        return wrapped_fun
 
     return wrapper
 
 
 def cleansup(fun) -> Callable:
-    """Method decorator to call `.cleanup()` after run."""
+    """Method decorator to call `.cleanup()` after run.
+
+    Can work with `sync` and `async` methods.
+    """
 
     @wraps(fun)
-    async def wrapped(self, *args, **kwargs) -> Optional[int]:
+    def wrapped(self, *args, **kwargs) -> Optional[int]:
+        try:
+            return fun(self, *args, **kwargs)
+        finally:
+            self.cleanup()
+
+    @wraps(fun)
+    async def async_wrapped(self, *args, **kwargs) -> Optional[int]:
         try:
             return await fun(self, *args, **kwargs)
         finally:
@@ -54,7 +80,11 @@ def cleansup(fun) -> Callable:
 
     # mypy doesnt trust `@wraps` to give back a `__wrapped__` object so we
     # need to code defensively here
-    wrapping = getattr(wrapped, "__wrapped__", None)
+    wrapped_fun = (
+        async_wrapped
+        if inspect.iscoroutinefunction(fun)
+        else wrapped)
+    wrapping = getattr(wrapped_fun, "__wrapped__", None)
     if wrapping:
         setattr(wrapping, "__cleansup__", True)
-    return wrapped
+    return wrapped_fun
